+++ GamePortal/GamePortal/Program.cs
@@ -1,8 +1,26 @@
+using GamePortal.Models;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
-using GamePortal.Models;
+using Serilog;
+using Serilog.Events;
+
+// Создание логов для пользователей
+Log.Logger = new LoggerConfiguration()
+    .MinimumLevel.Information()
+    // Чтобы в лог файл сейчас попадала только тнформация о создании пользователя и важные ошибки
+    .MinimumLevel.Override("Microsoft", LogEventLevel.Warning)
+    .MinimumLevel.Override("System", LogEventLevel.Warning)
+    .WriteTo.Console()
+    //
+    .WriteTo.File(
+    "Logs/LogUser.txt",
+    rollingInterval: RollingInterval.Day,
+    outputTemplate: "{Timestamp:yyyy-MM-dd HH:mm:ss} [{Level}] {Message}{NewLine}{Exception}")
+    .CreateLogger();
+    
 
 var builder = WebApplication.CreateBuilder(args);
+builder.Host.UseSerilog();
 
 var dbProvider = builder.Configuration["DbProvider"];
 
@@ -28,6 +46,7 @@ builder.Services.AddDefaultIdentity<IdentityUser>(options =>
 {
     options.SignIn.RequireConfirmedAccount = false;
 })
+.AddRoles<IdentityRole>() // Добавление для создания роли администратора
 .AddEntityFrameworkStores<ApplicationDbContext>();
 
 builder.Services.AddRazorPages();
@@ -50,5 +69,42 @@ app.UseAuthentication();
 app.UseAuthorization();
 
 app.MapRazorPages();
+
+// === Создание Администратора сайта ===
+using (var scope = app.Services.CreateScope())
+{
+    var services = scope.ServiceProvider;
+
+    var roleManager = services.GetRequiredService<RoleManager<IdentityRole>>();
+    var userManager = services.GetRequiredService<UserManager<IdentityUser>>();
+
+    // Создаем роль Admin если ее нет
+    if (!await roleManager.RoleExistsAsync("Admin"))
+    {
+        await roleManager.CreateAsync(new IdentityRole("Admin"));
+    }
+
+    // Проверяем существует ли пользователь
+    var adminEmail = "admin@example.com";
+    var adminUser = await userManager.FindByEmailAsync(adminEmail);
+
+    if (adminUser == null)
+    {
+        var newAdmin = new IdentityUser
+        {
+            UserName = adminEmail,
+            Email = adminEmail,
+            EmailConfirmed = true
+        };
+
+        var result = await userManager.CreateAsync(newAdmin, "Admin123!");
+
+        if (result.Succeeded)
+        {
+            await userManager.AddToRoleAsync(newAdmin, "Admin");
+            Log.Information("Создан администратор {Email}", adminEmail); // Информация о создании админа в логфайле
+        }
+    }
+}
 
 app.Run();