+++ GamePortal/GamePortal/CHANGELOG.md
@@ -6,3 +6,14 @@
 - Game purchase system
 - Game cards layout
 - Online games section( one test game)
+
+## [1.1.0] - 2026-02-17
+### Added
+- Role "Admin"
+- Automatic admin creation at startup
+- Access restriction for Edit/Delete pages
+- Serilog logging for user registration
+- Log level override for Microsoft and System
+
+### Changed
+- Improved authorization architecture